Moving dispute reason definitions to constants file for consistency
