+++ src/applications/dispute-debt/pages/disputeReason.js
@@ -1,8 +1,6 @@
 import VaRadioField from 'platform/forms-system/src/js/web-component-fields/VaRadioField';
 import DebtTitle from '../components/DebtTitle';
-
-const EXISTENCE = `I don't think I owe this debt to VA`;
-const AMOUNT = `I don't think the amount is correct on this debt`;
+import { DISPUTE_REASONS } from '../constants';
 
 const disputeReason = {
   uiSchema: {
@@ -15,10 +13,8 @@ const disputeReason = {
           'ui:required': () => true,
           'ui:options': {
             labels: {
-              // eslint-disable-next-line object-shorthand
-              EXISTENCE: EXISTENCE,
-              // eslint-disable-next-line object-shorthand
-              AMOUNT: AMOUNT,
+              EXISTENCE: DISPUTE_REASONS.EXISTENCE,
+              AMOUNT: DISPUTE_REASONS.AMOUNT,
             },
           },
           'ui:errorMessages': {
@@ -38,7 +34,7 @@ const disputeReason = {
           properties: {
             disputeReason: {
               type: 'string',
-              enum: [EXISTENCE, AMOUNT],
+              enum: [DISPUTE_REASONS.EXISTENCE, DISPUTE_REASONS.AMOUNT],
             },
           },
           required: ['disputeReason'],
